Move spawning background tasks to ffi module
